Created item object
i created a new file for creating  a random pos so i can use it in multiple files and when i decide to add more logical it will be implemented towards all of them. I also added items for now i have water and food that spawn in and wanna add logic so dinos start moving towards.