updating movement and states
Created new files for states/types/emotional states. When the dinos are in wander state they actually wander. The rule set i used is working so carnivores are changing to a chase state when a herbivore is in range. They change to a sleep state when low on health/food/water.